% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/china_water.R
\name{china_water}
\alias{china_water}
\title{china_water}
\usage{
china_water(type = "reservoir", outdir = ".", timestamp = TRUE)
}
\arguments{
\item{type}{reservoir, river, rain}
}
\description{
Get China realtime water data.
}
\examples{
\dontrun{
d_sk    <- china_water('reservoir')
d_river <- china_water('river')
d_rain  <- china_water('rain')
}
}
