% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curl_main.R
\name{curl_main}
\alias{curl_main}
\title{Main function for crawling China flow data}
\usage{
curl_main(date, outdir = ".", prefix = "", FUN = NULL)
}
\description{
Main function for crawling China flow data
}
\examples{
\dontrun{
library(lubridate)
dates <- seq(ymd("1946-01-01"), ymd("2019-04-25"), by = "day") #\%>\% format()
dates <- {dates + dhours(c(8))} \%>\% set_names(format_POSIXt(.))
l_ply(dates[1:8], download, outdir = "hubei", prefix = "hubei",
     FUN = curl_hubei)
}
}
