% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_cookie.R
\name{GET2}
\alias{GET2}
\alias{POST2}
\title{request url with cookiefile}
\usage{
GET2(..., cookiefile = "cookies.txt")

POST2(..., cookiefile = "cookies.txt")
}
\arguments{
\item{...}{other parameters will be passed to [GET()] or [POST()]}

\item{cookiefile}{path of cookie file

set_config() works globaly, but will not return cookie to reponse object.
Reversely, set_cookies() works locally, and return cookie to response object.}
}
\description{
request url with cookiefile
}
