% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_mete2000.R
\name{get_mete2000_stations}
\alias{get_mete2000_stations}
\title{get_mete2000_stations
It has been saved in \code{data('st_hourly')}.}
\usage{
get_mete2000_stations()
}
\description{
get_mete2000_stations
It has been saved in \code{data('st_hourly')}.
}
