% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curl_history.R, R/his_.R, R/his_HuNan.R
\name{curl_history}
\alias{curl_history}
\alias{his_}
\alias{his_HuBei}
\alias{his_yellowRiver}
\alias{his_HuNan}
\title{Main function for crawling China flow data}
\usage{
curl_history(date, outdir = ".", prefix = "", FUN = NULL)

his_HuBei(date, ...)

his_yellowRiver(date, ...)

his_HuNan(date = Sys.Date(), ...)
}
\arguments{
\item{date}{Date object, of querying}

\item{outdir}{directory of output files}

\item{prefix}{character, prefix of output file name}

\item{FUN}{function}

\item{...}{ignored.}
}
\description{
Main function for crawling China flow data

his_
}
\examples{
\dontrun{
library(lubridate)
dates <- seq(ymd("1946-01-01"), ymd("2019-04-25"), by = "day") #\%>\% format()
dates <- {dates + dhours(c(8))} \%>\% set_names(format_POSIXt(.))
plyr::l_ply(dates[1:8], curl_history, outdir = "hubei", prefix = "hubei",
     FUN = his_HuBei)
}

}
\references{
[1] http://113.57.190.228:8001/web/Report/RiverReport#

[2] HuNan, http://61.187.56.156/wap/index_sq.asp
}
