% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/his_.R
\name{his_hunan_type}
\alias{his_hunan_type}
\title{Hunan}
\usage{
his_hunan_type(date = Sys.Date(), type = "hnsq", ...)
}
\arguments{
\item{type}{hnsq: 全省水情日报
zykz: 控制站日报表
dxsk: 大型水库}
}
\description{
basins:
2: 湘江
3: 资水
4: 沅水
5: 澧水流域
6: 长江
7: 洞庭湖
}
\references{
http://61.187.56.156/wap/index_sq.asp
}
