% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{history}
\alias{history}
\title{Main function for crawling China flow data}
\usage{
history(date, outdir = ".", prefix = "", FUN = NULL)
}
\description{
Main function for crawling China flow data
}
\examples{
\dontrun{
library(lubridate)
dates <- seq(ymd("1946-01-01"), ymd("2019-04-25"), by = "day") #\%>\% format()
dates <- {dates + dhours(c(8))} \%>\% set_names(format_POSIXt(.))
plyr::l_ply(dates[1:8], history, outdir = "hubei", prefix = "hubei",
     FUN = his_hubei)
}

}
