% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{html_inputs}
\alias{html_inputs}
\title{html_inputs}
\usage{
html_inputs(p, xpath = "//input")
}
\arguments{
\item{p}{requested object from httr}

\item{xpath}{A string containing a xpath (1.0) expression.}
}
\description{
html_inputs
}
