% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url2params.R
\name{params2url}
\alias{params2url}
\title{constuct URL based on query parameters}
\usage{
params2url(url, params)
}
\arguments{
\item{url}{the url of the page to retrieve}

\item{params}{list}
}
\description{
constuct URL based on query parameters
}
