% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcurlFlow-package.R
\docType{package}
\name{RcurlFlow-package}
\alias{RcurlFlow}
\alias{RcurlFlow-package}
\title{RcurlFlow: Crawling China flow data}
\description{
Crawling China flow data.
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com}

}
\keyword{internal}
