% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curl_realtime.R
\name{RETRY2}
\alias{RETRY2}
\title{RETRY2}
\usage{
RETRY2(FUN, ..., times = 3)
}
\arguments{
\item{FUN}{function}

\item{...}{others to \code{FUN}}

\item{times}{max trying times}
}
\description{
RETRY2
}
