% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_GuangXi.R
\name{rt_GuangXi}
\alias{rt_GuangXi}
\title{rt_GuangXi}
\usage{
rt_GuangXi(outdir = ".", timestamp = TRUE, ...)
}
\arguments{
\item{...}{other parameters to [GET2()] or [POST2()]}
}
\description{
rt_GuangXi
}
