% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_mete2000.R
\name{rt_mete2000}
\alias{rt_mete2000}
\title{rt_mete2000}
\usage{
rt_mete2000(stations = NULL, date_begin, date_end, type = c("wap",
  "api"), outdir = ".", ...)
}
\arguments{
\item{date_begin}{POSIXct obj}

\item{date_end}{as date_begin}

\item{limits}{process how many sites}
}
\description{
This function only retrieves least 7 days meteorological data.
}
\examples{
info <- jsonlite::read_json('user_cma.json')
login_cma(info$user, info$pwd)
rt_mete2000()

}
