% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{stringToRaw}
\alias{stringToRaw}
\title{convert Raw format string into the real raw format variable}
\usage{
stringToRaw(str)
}
\arguments{
\item{raw}{format string}
}
\value{
raw vector
}
\description{
convert Raw format string into the real raw format variable
}
\examples{
\dontrun{
key_p <- "EB2A38568661887FA180BDDB5CABD5F21C7BFD59C090CB2D245A87AC253062882729
293E5506350508E7F9AA3BB77F4333231490F915F6D63C55FE2F08A49B353F444AD3993CACC02D
B784ABBB8E42A9B1BBFFFB38BE18D78E87A0E41B9B8F73A928EE0CCEE1F6739884B9777E4FE9E88A
1BBE495927AC4A799B3181D6442443"
stringToRaw(key_p)
}
}
