% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_stop.R
\name{task_runnow}
\alias{task_runnow}
\title{Immediately run a specific task available in the Windows task scheduler.}
\usage{
task_runnow(taskname)
}
\arguments{
\item{taskname}{the name of the task to run. See the example.}
}
\value{
the system call to schtasks /Run
}
\description{
Immediately run a specific task available in the Windows task scheduler.
}
\examples{
\dontrun{
myscript <- quote(print(Sys.time()))

task_create(taskname = "myfancyscript", rscript = myscript, 
     schedule = "ONCE", starttime = format(Sys.time() + 10*60, "\%H:\%M"))

task_runnow("myfancyscript")
}
}
