% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_stop.R
\name{task_stop}
\alias{task_stop}
\alias{task_delete}
\title{Delete or stop a specific task which was scheduled in the Windows task scheduler.}
\usage{
task_stop(taskname)

task_delete(taskname)
}
\arguments{
\item{taskname}{the name of the task to delete. See the example.}
}
\value{
the system call to schtasks /Delete or /End
}
\description{
Delete a specific task which was scheduled in the Windows task scheduler.
}
\examples{
\dontrun{
myscript <- quote(print(Sys.time()))
task_create(taskname = "myfancyscript", rscript = myscript, 
schedule = "ONCE", starttime = format(Sys.time() + 10*60, "\%H:\%M"))

task_runnow("myfancyscript")
Sys.sleep(5)
task_stop("myfancyscript")

task_delete(taskname = "myfancyscript")
}
}
