% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_mete2000.R
\name{union_mete2000_files}
\alias{union_mete2000_files}
\title{union_mete2000_files}
\usage{
union_mete2000_files(outdir, timerange)
}
\description{
union_mete2000_files
}
\examples{
\dontrun{
union_mete2000_files("OUTPUT/", "[2019042818,2019050518]")
}

}
