% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ChinaWater.R
\name{rt_ChinaWater}
\alias{rt_ChinaWater}
\title{rt_ChinaWater}
\usage{
rt_ChinaWater(type = "reservoir", outdir = ".", timestamp = TRUE,
  ...)
}
\arguments{
\item{type}{reservoir, river, rain}
}
\description{
Get China realtime water data.
}
\examples{
\dontrun{
d_sk    <- rt_ChinaWater('reservoir')
d_river <- rt_ChinaWater('river')
d_rain  <- rt_ChinaWater('rain')
}
}
