% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ChinaWater.R
\name{rt_ChinaWater}
\alias{rt_ChinaWater}
\title{rt_ChinaWater}
\usage{
rt_ChinaWater(type = "reservoir", outdir = ".", timestamp = TRUE,
  ...)
}
\arguments{
\item{type}{reservoir, river, rain}

\item{outdir}{directory of output files}

\item{timestamp}{boolean, whether include a timestamp to returned data.frame}

\item{...}{others to \code{FUN}}
}
\description{
Get China curl_realtime water data.
}
\examples{
\dontrun{
d_sk    <- rt_ChinaWater('reservoir')
d_river <- rt_ChinaWater('river')
d_rain  <- rt_ChinaWater('rain')
}
}
