% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_mete2000.R
\name{rt_mete2000}
\alias{rt_mete2000}
\title{rt_mete2000}
\usage{
rt_mete2000(stations = NULL, date_begin = NULL,
  date_end = Sys.time(), type = c("wap", "api"), outdir = "OUTPUT",
  limits = NULL, times = 3, ...)
}
\arguments{
\item{date_begin}{POSIXct obj, or character in the format of "%Y%m%d%H",
e.g "2019050518".}

\item{date_end}{as date_begin}

\item{type}{"api" or "wap"}

\item{outdir}{directory of output files}

\item{limits}{If not null, only process limits sites group}

\item{times}{max trying times}

\item{...}{others to \code{FUN}}
}
\description{
This function only retrieves least 7 days meteorological data.
}
\examples{
\dontrun{
info <- jsonlite::read_json('user_cma.json')
login_cma(info$user, info$pwd)
d <- rt_mete2000(date_end = "2019050518")
}

}
