% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookie.R
\name{set_cookie}
\alias{set_cookie}
\alias{write_cookie}
\title{once cookie exist, it's impossible to update by set_cookie}
\usage{
set_cookie(cookiefile = "cookies.txt")

write_cookie(cookies, cookiefile = "cookies.txt")
}
\description{
set_config() works globaly, but will not return cookie to reponse object.
Reversely, set_cookies() works locally, and return cookie to response object.
}
